+++ packages/terra-table/CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased
 ### Added
 * Added preventDefault to SingleSelect table row
 * Add missing props-tables
+* Updated nightwatch test scripts
+* Removed nightwatch.config file
+
 
 1.0.0 - (June 28, 2017)
 ------------------